Move OB specific changes to setMetaData step
