+++ components/org.wso2.carbon.identity.application.authenticator.push/pom.xml
@@ -254,7 +254,6 @@
                             org.wso2.carbon.identity.application.authenticator.push.notification.handler,
                             org.wso2.carbon.identity.application.authenticator.push.notification.handler.impl,
                         </Export-Package>
-                        <DynamicImport-Package>*</DynamicImport-Package>
                     </instructions>
                 </configuration>
             </plugin>
