+++ components/org.wso2.carbon.identity.application.authenticator.push/src/main/java/org/wso2/carbon/identity/application/authenticator/push/notification/handler/FirebasePushNotificationSender.java
@@ -83,7 +83,7 @@ public class FirebasePushNotificationSender {
     public void sendPushNotification(String deviceId, String pushId, String message, String randomChallenge,
                                      String sessionDataKey, String username, String fullName, String organization,
                                      String serviceProviderName, String hostname, String userOS, String userBrowser,
-                                     String consentInfo)
+                                     String metadata)
             throws AuthenticationFailedException {
 
         try {
@@ -116,7 +116,7 @@ public class FirebasePushNotificationSender {
             pushNotificationData.put(PushAuthenticatorConstants.REQUEST_DEVICE_BROWSER, userBrowser);
             pushNotificationData.put(PushAuthenticatorConstants.REQUEST_DEVICE_OS, userOS);
             // OB specific change to send consent data as metadata in the notification request body
-            pushNotificationData.put(PushAuthenticatorConstants.METADATA, consentInfo);
+            pushNotificationData.put(PushAuthenticatorConstants.METADATA, metadata);
 
             JSONObject json = new JSONObject();
             json.put(PushAuthenticatorConstants.TO, pushId.trim());
